Move utils to extensions/common. Remove infrastructure for old location.
