+++ src/extensions/common/Utility.ts
