+++ src/extensions/devel/ElectrodeGeometryTest/ElectrodeGeometry.tsx
@@ -1,9 +1,9 @@
 import { norm } from 'mathjs'
 import React from 'react'
-import { getArrayMax, getArrayMin } from '../../../util/Utility'
 import CanvasWidget from '../../common/CanvasWidget/CanvasWidget'
 import { CanvasWidgetLayer, useLayer, useLayers } from '../../common/CanvasWidget/CanvasWidgetLayer'
 import { getHeight, getWidth, RectangularRegion } from '../../common/CanvasWidget/Geometry'
+import { getArrayMax, getArrayMin } from '../../common/Utility'
 import { AnimatedLayerState, handleAnimatedClick, paintAnimationLayer } from './AnimatedLayer'
 import { ClickHistoryState, handleClickTrail, paintClickLayer } from './ClickHistoryLayer'
 import { DragLayerState, paintDragLayer, setDragLayerStateFromProps, updateDragRegion } from './DragLayer'
