+++ src/extensions/spikeamplitudes/SpikeAmplitudesView/useSpikeAmplitudesData.ts
@@ -1,7 +1,7 @@
 import { useContext, useMemo } from "react"
-import { getArrayMax, getArrayMin } from '../../../util/Utility'
 import { createCalculationPool, HitherContext, HitherInterface } from "../../common/hither"
 import useFetchCache from "../../common/useFetchCache"
+import { getArrayMax, getArrayMin } from '../../common/Utility'
 
 export type SpikeAmplitudesData = {
     getSpikeAmplitudes: (unitId: number) => {timepoints: number[], amplitudes: number[], minAmp: number, maxAmp: number} | undefined | null
